Update README.md
This Project is a website for listing IDOs that allows users to view upcoming and active IDOs via a smart contract, vote for their favorite IDOs, and comment on them. The site includes a leaderboard that displays the IDOs with the most votes at the top. The site also has a user interface designed with Tailwind CSS that includes a header and footer, as well as hover effects on all elements.